soc: qcom: memory_dump_v2: use kmemleak_not_leak for mem dump's app table
Table (of type msm_dump_table) is virtual address for app's base
table of memdump, which is converted to physical address and
holds it in the memdump.table_phys. This memory is not intended
to be freed. When kmemleak scans, it can't deduce that the reference
we are holding is via physical address and not virtual address.
Hence, kmemleak will treat this as a leak as it can't find
a reference to this variable. Suppress this by using
kmemleak_not_leak API.

Signed-off-by: Lingutla Chandrasekhar <clingutla@codeaurora.org>
(cherry picked from commit efeed9cca468fb4261c0c407231c0b6fa55b0035)

Change-Id: I2ec603d4efc19c791ae588f42855ba49a8776918
Signed-off-by: Manoharan Vijaya Raghavan <mraghava@codeaurora.org>