OpenWRT: 136-clk-Add-safe-switch-hook.patch
Sometimes clocks can't accept their parent source turning off
while the source is reprogrammed to a different rate. Most
notably CPU clocks require a way to switch away from the current
PLL they're running on, reprogram that PLL to a new rate, and
then switch back to the PLL with the new rate once they're done.
Add a hook that drivers can implement allowing them to return a
'safe parent' that they can switch their parent to while the
upstream source is reprogrammed to support this.

Change-Id: I7f439c306912352200bc1225085d941f573fe0bf
Signed-off-by: Ram Chandra Jangir <rjangi@codeaurora.org>
Signed-off-by: Abhishek Sahu <absahu@codeaurora.org>