ipq4019: Added support for APPS image upgrade
In order to support APPS upgrade in addition to failsafe
sysupgrade, the logic has been changed such that the
bootconfig1 is considered default irrespective of the age
and bootconfig2 is a backup containing the same data
as bootconfig1. Proc entries are created based on
information in bootconfig1.

Change-Id: I2a306fd764f6c63231c3a60f7f39e504787d3450
Signed-off-by: G Dhivya <gdhivya@codeaurora.org>
(cherry picked from commit 7be88cbb0820211949d6aea14cb0e8db607a6314)