clk: qcom: add parent map for regmap mux
Currently the driver assumes the register configuration value
is identical to its index in the parent map. This patch adds
the parent map field in regmap mux clock node which contains
the mapping of parent index with actual register configuration
value. If regmap node contains this parent map then the
configuration value will be taken from this
parent map instead of simply writing the index value.

Change-Id: Ia0c52365b523061184c7a4fe4c42b01c7f925810
Signed-off-by: Abhishek Sahu <absahu@codeaurora.org>