clk: qcom: ipq806x: port NSS clock to latest clock framework
Some of NSS virtual clock operations were commented in the
current code so uncommented the same and done the required
changes to make it working with latest clock framework.

Now the clock source value is assigned with enumeration
instead of hash defines so changes need to be done for
getting safe parent index. This patch fetches the safe parent
index in probe function which will be used for safe parent
related operations.

Change-Id: I65391186c9090442d504f30d86a895f24f88e473
Signed-off-by: Abhishek Sahu <absahu@codeaurora.org>