clk: qcom: support for non-continuous CMD_RCGR and CFG_RCGR
The current driver hardcodes the RCG2 registers offsets. Some
of the RCG2’s registers offsets are not according to the default
registers offsets in which CMD and CFG registers are continuous.
Now, this patch adds cfg register offset in rcg2 clock node
which should be assigned by driver. This value is CFG_RCGR
actual offset with respect to CMD_RCGR minus by 4 so that no
change required in existing drivers.

Change-Id: Ifc4b872f8cac8fdb350a667bd6f96257038d6515
Signed-off-by: Abhishek Sahu <absahu@codeaurora.org>