sound: ipq: Fix to handle audio pop noise during bootup
When the sound card registration fails, the audio TLMM change is also
reverted. Due to this, the pins are seen to toggle causing pop noise
during kernel bootup. To avoid this, the pins are set to GPIO state and
moved to audio functionality only when the sound card registration is
successful.

Change-Id: I0a9183245a76aafcb2f0b04837c45620e2ca8d8b
Signed-off-by: Anusha Canchi Ramachandra Rao <anusha@codeaurora.org>