ipq807x: rproc: Added authentication support using tz
rproc expects the firmware file to be created as
<filename>.mdt for the ELF header and hash segment part.
<filename>.b<nn> where <nn> is the ELF segment number.
the <filename>.b<nn> files are expected to have the segment
contents for each PHDR in the mdt file.

rproc loads the .mdt file and and gets authentication from
TZ. After successful authentication, it loads the .b<nn> files
and issues a request to TZ for bringing q6 out of reset.

Change-Id: I392db15c3f65affcecc1f70eb157ac9791210ac6
Signed-off-by: Manoharan Vijaya Raghavan <mraghava@codeaurora.org>