video: fbdev: support for mmap DMA coherent memory
The deferred IO overrides the framebuffer driver mmap function
with its own mmap function. Some of the driver framebuffer
memory is DMA coherent for which caching needs to be disabled.
This patch adds the flag FBINFO_DMAFB and set the appropriate
page protection for disabling caching if this flag is set.

Change-Id: I1157d9c7382ebcad3dde34654fc5a103acd5d81f
Signed-off-by: Abhishek Sahu <absahu@codeaurora.org>