ipq807x: Added the swiotlb option in 64-bit kernel command line
By default swiotlb allocates 64 MB of memory for bounce buffer.
In the current swiotlb implementation, if the DMA address used
is in addressable memory limit, swiotlb doesn't do any extra
processing. In ipq807x h/w all the peripherals use 32-bit
address in the dma transfers and the address ranges fall in the
addressable memory limit. Hence the swiotlb bounce buffers
can be reduced so that the 64MB can be freed for kernel usage.

Change-Id: Icea0adf7d2bf3e0cfbacf5d34f7b83de58160bf7
Signed-off-by: Jaiganesh Narayanan <njaigane@codeaurora.org>